Remove priors.py and introduce imfs.py
